% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amta-Zs.R
\name{amatz}
\alias{amatz}
\title{Adaptive multi-trait association test using GWAS summary data}
\usage{
amatz(Z, Sig, rho = 0:10/10)
}
\arguments{
\item{Z}{summary Z-statistics across multiple traits}

\item{Sig}{the estimated marginal trait correlation matrix}

\item{rho}{weights for the SZ}
}
\value{
\describe{
  \item{p.value}{ p-values for AZ, SZ2, and SZ }
  \item{pval}{ the list of all p-values }
  \item{rho.est}{ estimated optimal \eqn{\rho} value }
}
}
\description{
We consider a Z-stat based adaptive test (AZ) by weighting sum of Z-stats (SZ) and the sum of squared Z-stats (SZ2)
}
\references{
Guo,B. and Wu,B. (2018) Integrate multiple traits to detect novel disease-gene association using GWAS summary data with an adaptive test approach. \emph{Bioinformatics}, under revision.
}
