% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta.example}
\alias{beta.example}
\title{Example3: the genetic effect estimates and their standard errors}
\format{A list with 3 sublists:
\describe{
  \item{Beta}{a list of genetic effect estimates, each sublist is \eqn{m \times K} matrix for each study}
  \item{Beta.se}{a list of standard errors of genetic effect estimates, each sublist is \eqn{m \times K} matrix for each study}
  \item{R}{a SNP correlation matrix for the union of SNPs appearing in all the studies}}}
\usage{
data(beta.example)
}
\description{
An example of calculating U and V for MTAR given the genetic effect estimates and their standard errors.
}
\keyword{datasets}
