% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emats.R
\name{emats}
\alias{emats}
\title{Principal component based adaptive multi-trait association test using GWAS summary data}
\usage{
emats(Z, Sig, rho = 0:5/5)
}
\arguments{
\item{Z}{summary Z-statistics across multiple traits}

\item{Sig}{the estimated marginal trait correlation matrix}

\item{rho}{sequence of weights assigned to the 1-DF test}
}
\value{
\describe{
  \item{p.value}{ p-value for the adaptive test }
  \item{pvals}{ the list of all p-values }
  \item{rho.est}{ estimated optimal \eqn{\rho} value }
}
}
\description{
We show that using the summary statistics, we can compute the PC to summarize the multiple traits,
and then construct the 1-DF test statistic. The omnibus K-DF chi-square test is generally robust and powerful.
We then define the adaptive test as the minium p-values of weighted sums of 1-DF (\eqn{\rho}) and K-DF (\eqn{1-\rho}) tests.
Efficient algorithms are developed to compute the analytical p-values for all tests.
We use the LD score regression (see GCvr() function) to accurately estimate the marginal trait correlation using GWAS summary data.
}
\references{
Bulik-Sullivan B et al. (2015) An atlas of genetic correlations across human diseases and traits. \emph{Nature Genetics}, 47(11):1236--41.

Guo,B. and Wu,B. (2017) Principal component based adaptive association test of multiple traits using GWAS summary statistics. tech rep.
}
