% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katpval.R
\name{KATpval}
\alias{KATpval}
\title{Compute the tail probability of weighted sum of 1-DF chi-square rvs}
\usage{
KATpval(Q.all, lambda, acc = 0.01, lim = 1e+07)
}
\arguments{
\item{Q.all}{test statistics}

\item{lambda}{mixing coefficients}

\item{acc}{relative error bound}

\item{lim}{maximum number of integration terms}
}
\value{
tail probability
}
\description{
Use Davies' method and apply a relative error bound (avoid redundant computations): efficient and accurate.
}
\references{
Wu,B., Guan,W., Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. \emph{Annals of human genetics}, 80(2), 123-135.

Guo,B. and Wu,B. (2017) Principal component based adaptive association test of multiple traits using GWAS summary statistics. tech rep.
}
