% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amta-Zs.R
\name{Lamatz}
\alias{Lamatz}
\title{Compute the list of significant SNPs using the GWAS summary data based adaptive multi-trait association tests}
\usage{
Lamatz(Z, Sig, rho = 0:10/10, alpha = 5e-08)
}
\arguments{
\item{Z}{matrix of summary Z-statistics (SNPs by traits)}

\item{Sig}{the estimated marginal trait correlation matrix}

\item{rho}{weights for the SZ}

\item{alpha}{desired genome-wide significance level (default to 5E-8)}
}
\value{
\describe{
  \item{idAZ}{ significant SNP list for AZ }
}
}
\description{
We compute the significant SNPs for the adaptive test (AZ) based on weighting the SZ and SZ2 tests.
}
\references{
Guo,B. and Wu,B. (2018) Integrate multiple traits to detect novel disease-gene association using GWAS summary data with an adaptive test approach. \emph{Bioinformatics}, under revision.
}
