% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtaZ-list.R
\name{Lemats}
\alias{Lemats}
\title{Compute the list of significant SNPs using the GWAS summary data based multi-trait association tests}
\usage{
Lemats(Z, Sig, rho = 0:5/5, alpha = 5e-08)
}
\arguments{
\item{Z}{matrix of summary Z-statistics (SNPs by traits)}

\item{Sig}{the estimated marginal trait correlation matrix}

\item{rho}{sequence of weights assigned to the 1-DF test}

\item{alpha}{desired genome-wide significance level (default to 5E-8)}
}
\value{
\describe{
  \item{idQ}{ significant SNP list for OT }
  \item{idE}{ significant SNP list for ET }
  \item{idA}{ significant SNP list for AT }
}
}
\description{
We compute the significant SNPs for three tests: omnibus chi-square test (OT); 1-DF PC test (ET); adaptive test (AT).
}
\references{
Guo,B. and Wu,B. (2018) Integrate multiple traits to detect novel disease-gene association using GWAS summary data with an adaptive test approach. \emph{Bioinformatics}, under revision.
}
