% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtaZ-list.R
\name{Lmats}
\alias{Lmats}
\title{Compute the list of significant SNPs using the GWAS summary data based multi-trait association tests}
\usage{
Lmats(Z, Sig, rho = 0:5/5, alpha = 5e-08)
}
\arguments{
\item{Z}{matrix of summary Z-statistics (SNPs by traits)}

\item{Sig}{the estimated marginal trait correlation matrix}

\item{rho}{sequence of weights assigned to the 1-DF test}

\item{alpha}{desired genome-wide significance level (default to 5E-8)}
}
\value{
\describe{
  \item{idQ}{ significant SNP list for omnibus chi-square test }
  \item{idB}{ significant SNP list for 1-DF PC test }
  \item{idA}{ significant SNP list for adaptive test }
}
}
\description{
We compute the significant SNPs for three tests: omnibus chi-square test; 1-DF PC test; adaptive test.
}
\references{
Guo,B. and Wu,B. (2017) Principal component based adaptive association test of multiple traits using GWAS summary statistics. tech rep.
}
