% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtaZ-list.R
\name{Lmatz}
\alias{Lmatz}
\title{Compute the list of significant SNPs using the GWAS summary data based multi-trait association tests}
\usage{
Lmatz(Z, Sig, alpha = 5e-08)
}
\arguments{
\item{Z}{matrix of summary Z-statistics (SNPs by traits)}

\item{Sig}{the estimated marginal trait correlation matrix}

\item{alpha}{desired genome-wide significance level (default to 5E-8)}
}
\value{
\describe{
  \item{idM}{ significant SNP list for minP }
  \item{idSZ}{ significant SNP list for SZ }
  \item{idSZ2}{ significant SNP list for SZ2 }
  \item{idOT}{ significant SNP list for OT }
  \item{idMU}{ significant SNP list for min(SZ2,OT) }
  \item{idm}{ SNPs significant at any marginal trait }
}
}
\description{
We compute the significant SNPs for (1) minimum marginal test p-value (minP); SZ test; SZ2 test;
and (2) omnibus chi-square test (OT); minimum of OT and SZ2 (MU); and significant at any marginal test.
}
\references{
Guo,B. and Wu,B. (2018) Integrate multiple traits to detect novel disease-gene association using GWAS summary data with an adaptive test approach. \emph{Bioinformatics}, under revision.
}
