% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{MLM.null}
\alias{MLM.null}
\title{Null multivariate linear regression model fitting}
\usage{
MLM.null(Ys, Xs = NULL)
}
\arguments{
\item{Ys}{matrix of outcomes (samples in rows and trait in columns)}

\item{Xs}{matrix of common covariates for all traits. Default to NULL for no covariates.}
}
\value{
\describe{
  \item{res}{ residual matrix (sample in rows) }
  \item{Ux}{ eigen vectors of projection matrix }
  \item{n,m,p}{ sample size and outcome/covariate dimensions }
}
}
\description{
All outcomes are regressed on the same set of covariates.
}
