% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{MQTAd}
\alias{MQTAd}
\title{Multiple quantitative trait association test with differing covariates}
\usage{
MQTAd(obj, G)
}
\arguments{
\item{obj}{fitted null model from MDM.null}

\item{G}{genotype vector}
}
\value{
\describe{
  \item{p.value}{ three association p-values: an ombinus m-DF Wald test; two 1-DF Wald tests assuming common effect or common scaled effect (see ref) }
  \item{coef}{ estimated variant regression coefficients for all traits}
}
}
\description{
Fast computation of genome-wide association test of multiple quantitative traits.
}
\references{
Wu,B. and Pankow,J.S. (2017) Fast and accurate genome-wide association test of multiple quantitative traits. tech report.
}
