% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta-acl.R
\name{MTA.ACL}
\alias{MTA.ACL}
\title{A multinomial regression model for testing multi-trait association}
\usage{
MTA.ACL(Ys, Xs, G)
}
\arguments{
\item{Ys}{matrix of multiple continuous traits to be tested}

\item{Xs}{covariates to be adjusted}

\item{G}{genotype vector}
}
\value{
\describe{
  \item{p.value}{ three p-values: an ombinus m-DF chi-square test; two 1-DF chi-square tests assuming common effect or common scaled effect (see ref) }
}
}
\description{
An ACL (adjacent-category-logit) model that accounts for population stratification and tests multi-trait association.
}
\references{
Wu,B. and Pankow,J.S. (2015) Statistical methods for association tests of multiple continuous traits in genome-wide association studies. \emph{Annals of human genetics}, 79(4), 282-293.
}
