% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeLin.R
\name{MTA.gee}
\alias{MTA.gee}
\title{GEE association test of multiple quantitative traits}
\usage{
MTA.gee(obj, G)
}
\arguments{
\item{obj}{fitted null model from MLM.null}

\item{G}{genotype vector}
}
\value{
\describe{
  \item{p.value}{ three association p-values: an ombinus m-DF chi-square test; two 1-DF chi-square tests assuming common effect or common scaled effect (see ref) }
  \item{U,V}{ score vector (U) and its asymptotic covariance matrix (V) }
}
}
\description{
An efficient implementation of genome-wide association test of multiple quantitative traits using GEE score tests.
}
\references{
He,Q., Avery,C.L. and Lin,D.Y. (2013) A general framework for association tests with multivariate traits in large-scale genomics studies. \emph{Genetic Epidemiology}, 37 (8), 759–767.

Wu,B. and Pankow,J.S. (2017) Fast and accurate genome-wide association test of multiple quantitative traits. tech report.
}
