% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{MTA.minp}
\alias{MTA.minp}
\title{Minimum p-value based multiple quantitative trait association test}
\usage{
MTA.minp(obj, G)
}
\arguments{
\item{obj}{fitted null model from MLM.null}

\item{G}{genotype vector}
}
\value{
\describe{
  \item{p.value}{ individual trait association test p-values }
  \item{Pbonf}{ Bonferroni corrected minimum p-value }
  \item{Pmin}{ Significance p-value of minimum p-value computed analytically based on multivariate normal dist }
}
}
\description{
We efficiently compute minimum test p-values (minP) across multiple quantitative traits;
and report two test p-values: 1) Bonferroni corrected minimum p-values (Pbonf) across traits.
2) Analytical significance p-value (Pmin) of minP using asymptotic multivariate normal integration.
See refs in the PGmvn() function.
}
\details{
Remarks: (1) Be cautious to interpret extreme Pmin: it is very hard to accurately compute extreme p-values.
(2) Generally Pmin is close to Pbonf at extreme values under moderate trait correlations.
(3) Under extreme trait correlations, Pmin can offer advantages compared to Pbonf.
(4) Note that theoretically we have minP \eqn{\le} Pmin \eqn{\le} Pbonf.
}
\references{
Conneely,K.N. and Boehnke,M. (2007) So many correlated tests, so little time! Rapid adjustment of P values for multiple correlated tests. Am. J. Hum. Genet. 81, 1158–1168.

Conneely,K.N. and Boehnke,M. (2010) Meta-analysis of genetic association studies and adjustment for multiple testing of correlated SNPs and traits. Genetic Epidemiology. 34:739-746.

Wu,B. (2017) MTAR: an R package for versatile genome-wide association test of multiple traits. tech report.
}
