% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTAR.example}
\alias{MTAR.example}
\title{PNPLA2 Gene}
\format{A list with 6 sublists:
\describe{
  \item{annotation}{some annotation information of gene PNPLA2}
  \item{U}{a list containing the summary statistics for each trait}
  \item{V}{a list containing the covariance matrix of summary statistics for each trait}
  \item{MAF}{the minor allele frequency of all the rare variants in this gene}
  \item{genetic_cor.trait}{the genetic correlation among traits}
  \item{zeta}{the overlapping estimation matrix, which is approximated by the sample correlation matrix of the scaled summary statistics over a large number of independent null common SNPs}}}
\usage{
data(MTAR.example)
}
\description{
PNPLA2 Gene Inofmation required by MTAR function.
}
\keyword{datasets}
