% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmats.R
\name{pmats}
\alias{pmats}
\title{Pleiotropy informed adaptive association test across multiple traits using summary statistics}
\usage{
pmats(Z, Sig, rho = 0:10/10)
}
\arguments{
\item{Z}{summary Z-statistics across multiple traits}

\item{Sig}{estimated trait correlation matrix}

\item{rho}{sequence of weights assigned to the PT. Default to 0:10/10}
}
\value{
\describe{
  \item{p.value}{ vector of p-values for: AT, OT and PT }
  \item{pvals}{ the list of all p-values }
  \item{rho.est}{ estimated optimal \eqn{\rho} value }
}
}
\description{
Multi-trait association test based on GWAS summary statistics under common genetic effect assumption
}
\details{
Under pleiotropy of common effects across traits, an 1-DF pleiotropy test (PT) can be constructed.
The omnibus K-DF chi-square test (OT) is generally robust and powerful.
We then define the adaptive test (AT) as the minium p-values of weighted sums of PT (\eqn{\rho}) and OT (\eqn{1-\rho}).
An efficient algorithm is developed to exactly compute the analytical p-value of AT.
We recommend using the LD score regression (Bulik-Sullivan et al.) to accurately estimate the trait correlation matrix.
}
\examples{
Sig = diag(3)
Sig[c(2:3,6, 4,7,8)] = c(0.2646,-0.2317, 0.7451)
Z = c(1,2,4)
pmats(Z,Sig)
}
\references{
Bulik-Sullivan BK  et al. (2015) LD Score regression distinguishes confounding from polygenicity in genome-wide association studies.
 \emph{Nature Genetics}, 47(3):291--295.

Bulik-Sullivan B et al. (2015) An atlas of genetic correlations across human diseases and traits. \emph{Nature Genetics}, 47(11):1236--1241.

Massoti M*, Guo B* and Wu B. (2018) Pleiotropy informed adaptive association test of multiple traits using GWAS summary data. \emph{Biometrics}, under revision (* contribute equallly).
}
