% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc-rlm1.R
\name{SHvr}
\alias{SHvr}
\title{A robust and efficient LD score regression (LDSC) for estimating the SNP heritability using GWAS summary data}
\usage{
SHvr(Z, r2, N, W = NULL)
}
\arguments{
\item{Z}{summary Z-statistics for M variants}

\item{r2}{average reference LD scores for M variants}

\item{N}{GWAS sample size for each variant (could be different across variants)}

\item{W}{variant weight}
}
\value{
\describe{
  \item{h2}{ SNP heritability }
  \item{v0}{ intercept in the LDSC, related to the genomic control (GC) parameter }
}
}
\description{
We follow the Bulik-Sullivan et al. (2015) approach but use a robust LS estimation without filtering large summary statistics.
The input r2 is the averaged reference LD scores (scaled by the total number of SNPs used to compute the LD scores).
The weight W is typically based on the LD scores computed using HapMap3 common SNPs. The purpose is to correct over-counting.
LDSC is a variance regression (VR) approach: regressing chi-square statistics on LD scores.
}
\references{
Bulik-Sullivan,B.K. et al. (2015) LD Score regression distinguishes confounding from polygenicity in genome-wide association
studies. Nature Genetics, 47, 291-295.

Guo,B. and Wu,B. (2018) Principal component based adaptive association test of multiple traits using GWAS summary statistics.
}
