% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varU.example}
\alias{varU.example}
\title{Example2: the summary statistics and their variance}
\format{A list with 3 sublists:
\describe{
  \item{U}{a list of summary statistics, each sublist is \eqn{m \times K} matrix for each study}
  \item{varU}{a list of variance of summary statistics, each sublist is \eqn{m \times K} matrix for each study}
  \item{R}{a SNP correlation matrix for the union of SNPs appearing in all the studies}}}
\usage{
data(varU.example)
}
\description{
An example of calculating U and V for MTAR given the summary statistics and its variance.
}
\keyword{datasets}
