% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeta.example}
\alias{zeta.example}
\title{Example4: the summary statistics of 737 common and null SNPs}
\format{A list with 1 sublist:
\describe{
  \item{Zscore}{a list, each sublist contains the Z-scores of 737 null and common SNPs for each trait}
  }}
\usage{
data(zeta.example)
}
\description{
An example of estimating matrix \eqn{\zeta} given the summary statistics information of null and common SNPs.
}
\keyword{datasets}
