% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta-Zs.R
\name{matz}
\alias{matz}
\title{Various multi-trait association tests using GWAS summary data}
\usage{
matz(Z, Sig)
}
\arguments{
\item{Z}{summary Z-statistics across multiple traits}

\item{Sig}{the estimated marginal trait correlation matrix}
}
\value{
vector of p-values for three tests:  minP,SZ,SZ2
}
\description{
We compute analytical p-values for three multi-trait tests: minimum marginal test p-value (minP), sum of Z-stats (SZ),
and sum of squared Z-stats (SZ2).
}
\references{
Massoti M*, Guo B* and Wu B. (2018) Pleiotropy informed adaptive association test of multiple traits using GWAS summary data. \emph{Biometrics}, to appear (* contribute equallly).
}
